Add highlight word count feature to chat messages
This commit introduces the ability to highlight a specific number of words in chat messages. The `highlight_word_count` property has been added to the `ChatMessage` and `ChatMessageType` interfaces, and the logic for applying the highlight styling based on this count has been implemented in the `ChatMessage` component. The `ChatTile` component has been updated to pass the `highlight_word_count` to the `ChatMessage` component. Additionally, the default value for `highlight_word_count` is set to 0 when sending or updating messages in the chat system. This feature enhances the visual feedback for chat interactions by emphasizing certain parts of the message.